Split gulpfile.js into separate files
